Remove check for whether channel is secure/insecure. Always prefer 'https_proxy' and fallback on 'http_proxy'
